Fixed behavior of addItemsToImport. Won't break import interface.
The way it was implemented, calling addItemsToImport on an ImportDecl with no explicit function import list would create said list, adding the specified elements and eliminating every other one. Reconsidering on the design decisions from the old API, I came to the conclusion that this would be a wrong behavior. We should keep from adding anything if the ImportDecl doesn't already specify a list of functions.